% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discharge.R
\name{LateralSubsurfaceDischarge}
\alias{LateralSubsurfaceDischarge}
\title{Lateral Subsurface Discharge}
\usage{
LateralSubsurfaceDischarge(psi, elevation, slope = NULL, aspect = NULL,
  soilmap, soiltypes, ...)
}
\arguments{
\item{psi}{soil presssure head. It is  \code{RasterBrick-class} object.}

\item{hsub}{surface water depth. It is  \code{RasterBrick-class} object.}

\item{elev}{}
}
\description{
Lateral Subsurface Discharge
}
\seealso{
\code{\link{terrain}}
}

