% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/soilpropmap.R
\name{SoilPropertyMap}
\alias{SoilPropertyMap}
\title{Craetion of a Physical Variable Map from a Catagory Map}
\usage{
SoilPropertyMap(x, map = "SoilMapFile", metacat = "SoilParFile",
  header = "Header", names = NULL, ...)
}
\arguments{
\item{x}{variable to map}

\item{map}{catagory input map, i.e. soiltype maps or related GEOtop keyword.}

\item{metacat}{list of dataframe containingcatagory definitions, e. g. the list of soil properties for each class or related GEOtop keyword.}

\item{header}{header sting using in \code{geotop.inpts} file. Default is \code{"Header"}}

\item{...}{further arguments for \code{\link{get.geotop.inpts.keyword.value}} such as \code{wpath} or \code{inpts.file}                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       s


wpath <- system.file("geotop_simulation",package="AnalyzeMoistureInSuperslab")
x <- "Alpha"
b <- SoilPropertyMap(x,wpath=wpath)

### BELOW a list of some keywords used for soil property variables
#HeaderSoilDz    =    "Dz"
#HeaderLateralHydrConductivity    =    "Kh"
#HeaderNormalHydrConductivity    =    "Kv"
#HeaderThetaRes    =    "vwc_r"
#HeaderWiltingPoint    =    "vwc_w"
#HeaderFieldCapacity    =    "vwc_fc"
#HeaderThetaSat    =    "vwc_s"
#HeaderAlpha    =    "alpha"
#HeaderN    =    "n"
#HeaderSpecificStorativity    =    "stor"
#HeaderKthSoilSolids    =    "Kth"
#HeaderCthSoilSolids    =    "Cth"
#HeaderSoilInitPres = "InitPsi"}
}
\description{
This functions creates physicalvariable maps from catagorical information, e.g. a map of hydraulic conductivity from the map of Soil Categories.
}

