% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/profile.R
\name{SoilVariableProfile}
\alias{SoilVariableProfile}
\title{Soil Variable weather Profile}
\usage{
SoilVariableProfile(x, points, names = NULL...)
}
\arguments{
\item{x}{soil variable \code{RasterBrick-class} object}

\item{points}{....}
}
\description{
Vertical or Slope-Normal Integration/Aggregation of soil water content
}
\examples{
data(PsiTheta)

 t <- 5
 map <- psi[[t]]
 yv <- (ymax(map)+ymin(map))/2
 points <- data.frame(x=c(32,41),y=yv,id=c("A","B"))

mapProfile <- SoilVariableProfile(map,points[,c("x","y")],names=as.character(points$id))

plot(mapProfile[,1],c(100:1),type="l")
lines(mapProfile[,1],c(100:1))
}

