% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SoilVariableProfile}
\alias{SoilVariableProfile}
\title{Soil Variable weather Profile}
\usage{
SoilVariableProfile(x, points, names = NULL...)
}
\arguments{
\item{points}{....}

\item{map}{soil variable \code{RasterBrick-class} object}
}
\description{
Vertical or Slope-Normal Integration/Aggregation of soil water content
}
\examples{
data(PsiTheta)

 t <- 5
 map <- psi[[t]]
 yv <- (ymax(map)+ymin(map)/2
 points <- data.frame(x=c(32,41),y=yv)
}

